🤏 small refactor (confirmation modularized) (moved invitelink module)
